requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
dj-database-url
gunicorn
whitenoise
python-dotenv
